+++ src/client.ts
@@ -2,18 +2,20 @@ import { SocketClient } from './socketAdapter/socketClient';
 import { createWebsocketClient } from './socketAdapter/websocketClient';
 import { parsePath } from './parsePath';
 import { createStore, Store } from './store';
-import { isObject, joinPath, mergeDiff } from './utils';
+import { deepClone, isObject, joinPath, mergeDiff } from './utils';
 import { Node, nodeify, traverseNode, unwrap } from './node';
 import { BatchedUpdate, createUpdateBatcher } from './updateBatcher';
+import { Plugin } from './plugin';
+import { createHooks, Hook } from './hooks';
 
 type Unsubscriber = () => void;
 
-type Meta = { [namespace: string]: any };
+type Meta = Node['meta'];
 
 export type ChainReference = {
 	get: (path: string) => ChainReference;
 	each: (callback: (ref: ChainReference, key: string) => void) => Unsubscriber;
-	set: (value: any) => ChainReference;
+	set: (value: any, meta?: Meta) => ChainReference;
 	delete: () => void;
 	on: (callback: (data: any, meta: Meta) => void) => Unsubscriber;
 	once: (callback: (data: any, meta: Meta) => void) => void;
@@ -23,16 +25,27 @@ type UpdateListener = {
 	[path: string]: ((data: any) => void)[];
 };
 
+export type ClientHooks = {
+	'client:set'?: Hook<{ path: string; value: any; meta: Meta }>;
+	'client:firstConnect'?: Hook<void>;
+	'client:reconnect'?: Hook<void>;
+	'client:disconnect'?: Hook<void>;
+};
+
+export type ClientPlugin = Plugin<ClientHooks>;
+
 export function SocketDBClient({
 	url,
 	store = createStore(),
 	socketClient,
 	updateInterval = 50,
+	plugins = [],
 }: {
 	url?: string;
 	store?: Store;
 	socketClient?: SocketClient;
 	updateInterval?: number;
+	plugins?: ClientPlugin[];
 } = {}) {
 	if (!url && !socketClient)
 		url =
@@ -43,6 +56,9 @@ export function SocketDBClient({
 
 	const subscribedPaths: string[] = [];
 	const updateListener: UpdateListener = {};
+	const hooks = createHooks<ClientHooks>();
+
+	registerPlugins(plugins);
 
 	const queueUpdate = createUpdateBatcher((diff) => {
 		socketClient.send('update', { data: diff });
@@ -59,10 +75,26 @@ export function SocketDBClient({
 					socketClient.send('subscribe', { path });
 				}
 			});
+			hooks.call('client:reconnect');
 			connectionLost = false;
+		} else {
+			hooks.call('client:firstConnect');
 		}
 	});
-	socketClient.onDisconnect(() => (connectionLost = true));
+	socketClient.onDisconnect(() => {
+		hooks.call('client:disconnect');
+		connectionLost = true;
+	});
+
+	function registerPlugins(plugins: ClientPlugin[]) {
+		plugins.forEach((plugin) => {
+			Object.entries(plugin.events).forEach(
+				([name, hook]: [keyof ClientHooks, ClientHooks[keyof ClientHooks]]) => {
+					hooks.register(name, hook);
+				}
+			);
+		});
+	}
 
 	function notifySubscriber(diff: Node) {
 		const listener: UpdateListener = { ...updateListener };
@@ -70,8 +102,7 @@ export function SocketDBClient({
 		// should not go through multiple loops
 		traverseNode(diff, (path, data) => {
 			if (listener?.[path]) {
-				let storedData = store.get(path);
-				storedData = mergeDiff(storedData, {}) as Node; // deep copy
+				let storedData = deepClone(store.get(path));
 				listener[path].forEach((listener) => listener(storedData));
 				delete listener[path];
 			}
@@ -81,7 +112,7 @@ export function SocketDBClient({
 				Object.keys(listener).forEach((subscribedPath) => {
 					if (subscribedPath.startsWith(path)) {
 						let storedData = store.get(subscribedPath);
-						storedData = mergeDiff(storedData, {}) as Node; // deep copy
+						storedData = deepClone(storedData);
 						listener[subscribedPath].forEach((listener) =>
 							listener(storedData)
 						);
@@ -195,12 +226,32 @@ export function SocketDBClient({
 					removeSocketPathSubscription(wildcardPath);
 				};
 			},
-			set(value) {
+			set(value, meta) {
 				if (!connectionLost) {
-					const diff = store.put(creatUpdate(path, nodeify(value)));
-					if (diff && Object.keys(diff).length > 0)
-						queueUpdate({ type: 'change', data: diff });
-					notifySubscriber(diff);
+					let clonedValue = value;
+					let clonedMeta = meta;
+					// deep clone only if we have hooks, store.put already does a deep clone
+					if (hooks.count('client:set') > 0) {
+						clonedValue = deepClone(value);
+						clonedMeta = deepClone(meta);
+					}
+					hooks
+						.call(
+							'client:set',
+							({ path, value, meta }) => {
+								const node = nodeify(value);
+								if (meta) node.meta = meta;
+								const update = creatUpdate(path, node);
+								const diff = store.put(update);
+								if (diff && Object.keys(diff).length > 0)
+									queueUpdate({ type: 'change', data: diff });
+								notifySubscriber(diff);
+							},
+							{ path, value: clonedValue, meta }
+						)
+						.catch((e) => {
+							console.log(e);
+						});
 				}
 				return this;
 			},